Move poem parser service to separate service
