+++ src/main/java/com/kbui/ceres/service/crawler/PoemFetcherImpl.java
@@ -1,12 +1,7 @@
 package com.kbui.ceres.service.crawler;
 
-import com.kbui.ceres.exception.crawler.PoemParsingException;
-import com.kbui.ceres.repository.PoemRepository;
 import com.kbui.ceres.repository.PoetRepository;
-import com.kbui.ceres.repository.models.PoemEntity;
 import com.kbui.ceres.repository.models.Poet;
-import com.kbui.ceres.scheduler.PoemCrawlerTask;
-import com.kbui.ceres.service.crawler.entity.Poem;
 import com.kbui.ceres.service.crawler.parser.PoemParserService;
 import io.vertx.core.CompositeFuture;
 import io.vertx.core.Future;
@@ -21,14 +16,14 @@ import org.springframework.stereotype.Component;
 
 @Component
 public class PoemFetcherImpl implements PoemFetcher {
-  private static final Logger log = LoggerFactory.getLogger(PoemCrawlerTask.class);
+  private static final Logger log = LoggerFactory.getLogger(PoemFetcherImpl.class);
   private Poet poet;
   private String poetName;
   private String baseUrl;
 
   @Autowired private PoetRepository poetRepository;
 
-  @Autowired private PoemRepository poemRepository;
+  @Autowired private PoemParserService poemParserService;
 
   public String getPoetName() {
     return poetName;
@@ -46,7 +41,10 @@ public class PoemFetcherImpl implements PoemFetcher {
         .compose(
             links -> {
               var futures =
-                  Arrays.asList(links.stream().map(this::getAndSavePoem).toArray(Future[]::new));
+                  Arrays.asList(
+                      links.stream()
+                          .map(link -> poemParserService.getAndSavePoem(link, this.poetName))
+                          .toArray(Future[]::new));
               return Future.future(
                   (future) ->
                       CompositeFuture.all(futures).onComplete(r -> future.complete(r.succeeded())));
@@ -127,66 +125,5 @@ public class PoemFetcherImpl implements PoemFetcher {
             ex ->
                 log.error(
                     String.format("Failed to fetch links from poet page: %s", ex.getMessage())));
-  }
-
-  private Future<PoemEntity> getAndSavePoem(String link) {
-    return poemRepository
-        .get(link)
-        .compose(this::updatePoem)
-        .recover(ex -> this.createPoem(link))
-        .onComplete(
-            r -> {
-              if (r.result().getId() != null) {
-                log.info("Link fetching done: {}", link);
-              }
-            });
-  }
-
-  private Future<PoemEntity> createPoem(String link) {
-    return getPoem(link)
-        .compose(
-            poem ->
-                poemRepository.save(
-                    PoemEntity.builder()
-                        .name(poem.getTitle())
-                        .url(link)
-                        .content(poem.getContent())
-                        .author(getPoetName())
-                        .build()))
-        .recover(
-            e -> {
-              log.error("Failed to fetch poem...Gonna try latter in the next task.", e);
-              return Future.succeededFuture(PoemEntity.empty());
-            });
-  }
-
-  private Future<PoemEntity> updatePoem(PoemEntity poemEntity) {
-    if (poemEntity.getContent().size() == 0
-        || poemEntity.getName().isEmpty()) { // in case we were detected by bot.
-      return getPoem(poemEntity.getUrl())
-          .compose(p -> poemRepository.update(poemEntity.getUrl(), p.getContent(), p.getTitle()))
-          .recover(
-              ex -> {
-                log.error(
-                    "Failed to update failed fetch attempt...Gonna try latter in the next task.");
-                return Future.succeededFuture(PoemEntity.empty());
-              });
-    } else {
-      return Future.succeededFuture(poemEntity);
-    }
-  }
-
-  private Future<Poem> getPoem(String link) {
-    return BaseCrawler.fetchPage(link)
-        .flatMap(
-            page -> {
-              try {
-                var poem = PoemParserService.parse(page);
-                return Future.succeededFuture(poem);
-              } catch (PoemParsingException ex) {
-                log.error("Failed to fetch poem in this link: {}", link);
-                return Future.failedFuture(ex);
-              }
-            });
   }
 }
